+++ plugins/changelayer.cpp
@@ -1,25 +1,24 @@
 // changelayer plugin
 // allows changing the material type of geological layers
 
-// some headers required for a plugin. Nothing special, just the basics.
+#include "Console.h"
 #include "Core.h"
-#include <Console.h>
-#include <Export.h>
-#include <PluginManager.h>
-
-// DF data structure definition headers
 #include "DataDefs.h"
-#include "modules/Maps.h"
-#include "modules/Materials.h"
-#include "modules/MapCache.h"
-#include "modules/Gui.h"
-
+#include "Export.h"
+#include "PluginManager.h"
 #include "TileTypes.h"
 
+#include "modules/Gui.h"
+#include "modules/MapCache.h"
+#include "modules/Maps.h"
+#include "modules/Materials.h"
+
+// DF data structure definition headers
+#include "df/region_map_entry.h"
+#include "df/world.h"
 #include "df/world_data.h"
 #include "df/world_geo_biome.h"
 #include "df/world_geo_layer.h"
-#include "df/region_map_entry.h"
 
 using namespace DFHack;
 using namespace df::enums;
@@ -228,7 +227,7 @@ command_result changelayer (color_ostream &out, std::vector <std::string> & para
     {
         if(verbose)
             out << "---Biome: " << i;
-        if(!all_biomes && i!=biome)
+        if(!all_biomes && uint32_t(i)!=biome)
         {
             if(verbose)
                 out << "-skipping" << endl;
@@ -258,7 +257,7 @@ command_result changelayer (color_ostream &out, std::vector <std::string> & para
             out << "geoindex: " << geoindex << endl;
 
         bool skip = false;
-        for(int g=0; g<v_geoprocessed.size(); g++)
+        for(int g=0; size_t(g)<v_geoprocessed.size(); g++)
         {
             if(v_geoprocessed.at(g)==geoindex)
             {
